old-django-app/backend/requirements.txt
asgiref==3.2.7
autopep8==1.5.1
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Jinja2==2.11.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1